Reducing opentracing version and reverting segregation by interface
