+++ tracing/build.gradle
@@ -3,8 +3,8 @@ dependencies {
     compileOnly group: 'com.newrelic.agent.java', name: 'newrelic-api', version: '4.4.0'
 
     //Datadog
-    compileOnly group: 'io.opentracing', name: 'opentracing-api', version: "0.32.0"
-    compileOnly group: 'io.opentracing', name: 'opentracing-util', version: "0.32.0"
+    compileOnly group: 'io.opentracing', name: 'opentracing-api', version: "0.31.0"
+    compileOnly group: 'io.opentracing', name: 'opentracing-util', version: "0.31.0"
     compileOnly group: 'com.datadoghq', name: 'dd-trace-ot', version: '0.26.0'
     compileOnly group: 'com.datadoghq', name: 'dd-trace-api', version: '0.26.0'
     compileOnly group: 'com.datadoghq', name: 'java-dogstatsd-client', version: '2.7'
