+++ tracing/src/main/kotlin/br/com/guiabolso/tracing/engine/datadog/DatadogStatsDTracer.kt
@@ -1,15 +1,13 @@
 package br.com.guiabolso.tracing.engine.datadog
 
-import br.com.guiabolso.tracing.engine.TimeRecorderTracerEngine
 import com.timgroup.statsd.NonBlockingStatsDClient
-import io.opentracing.Tracer
 
 
 class DatadogStatsDTracer(
         prefix: String,
         host: String,
         port: Int
-) : DatadogTracer(), TimeRecorderTracerEngine<Tracer.SpanBuilder> {
+) : DatadogTracer() {
 
     private val statsDClient = NonBlockingStatsDClient(prefix, host, port)
