+++ tracing/src/main/kotlin/br/com/guiabolso/tracing/engine/newrelic/NewRelicTracer.kt
@@ -1,6 +1,6 @@
 package br.com.guiabolso.tracing.engine.newrelic
 
-import br.com.guiabolso.tracing.engine.TimeRecorderTracerEngine
+import br.com.guiabolso.tracing.engine.TracerEngine
 import com.newrelic.api.agent.NewRelic
 import com.newrelic.api.agent.NewRelic.addCustomParameter
 import com.newrelic.api.agent.NewRelic.noticeError
@@ -9,7 +9,7 @@ import com.newrelic.api.agent.NewRelic.setTransactionName
 import com.newrelic.api.agent.Token
 import java.io.Closeable
 
-class NewRelicTracer : TimeRecorderTracerEngine<Token> {
+class NewRelicTracer : TracerEngine<Token> {
     override fun setOperationName(name: String) {
         setTransactionName(null, name)
     }
