+++ tracing/src/main/kotlin/br/com/guiabolso/tracing/engine/slf4j/Slf4JTracer.kt
@@ -1,11 +1,11 @@
 package br.com.guiabolso.tracing.engine.slf4j
 
-import br.com.guiabolso.tracing.engine.TimeRecorderTracerEngine
+import br.com.guiabolso.tracing.engine.TracerEngine
 import org.slf4j.LoggerFactory
 import org.slf4j.MDC
 import java.io.Closeable
 
-class Slf4JTracer : TimeRecorderTracerEngine<Map<String, String?>> {
+class Slf4JTracer : TracerEngine<Map<String, String?>> {
 
     override fun setOperationName(name: String) {
         addProperty("Operation", name)
