+++ tracing/src/main/kotlin/br/com/guiabolso/tracing/utils/EnvironmentUtils.kt
@@ -1,0 +1,10 @@
+package br.com.guiabolso.tracing.utils
+
+object EnvironmentUtils {
+
+    @JvmStatic
+    fun getProperty(key: String, default: String) = System.getenv(key) ?: default
+
+    @JvmStatic
+    fun getProperty(key: String, default: Int) = System.getenv(key)?.let { it.toInt() } ?: default
+}
